+++ lib/AppInfo/Application.php
@@ -1,11 +1,14 @@
 <?php
+
 declare(strict_types=1);
+
 /**
  * @copyright Copyright (c) 2016, Roeland Jago Douma <roeland@famdouma.nl>
  *
  * @author Roeland Jago Douma <roeland@famdouma.nl>
+ * @author Richard Steinmetz <richard@steinmetz.cloud>
  *
- * @license GNU AGPL version 3 or any later version
+ * @license AGPL-3.0-or-later
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Affero General Public License as
@@ -21,32 +24,27 @@ declare(strict_types=1);
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  *
  */
+
 namespace OCA\PreviewGenerator\AppInfo;
 
-use OCA\PreviewGenerator\Watcher;
+use OCA\PreviewGenerator\Listeners\PostWriteListener;
 use OCP\AppFramework\App;
-use OCP\AppFramework\IAppContainer;
-use OCP\Files\IRootFolder;
-use OCP\Files\Node;
+use OCP\AppFramework\Bootstrap\IBootContext;
+use OCP\AppFramework\Bootstrap\IBootstrap;
+use OCP\AppFramework\Bootstrap\IRegistrationContext;
+use OCP\Files\Events\Node\NodeWrittenEvent;
 
-class Application extends App {
-
-	const APPNAME='previewgenerator';
+class Application extends App implements IBootstrap {
+	public const APP_ID = 'previewgenerator';
 
 	public function __construct() {
-		parent::__construct(self::APPNAME);
-
-		$container = $this->getContainer();
-		$this->connectWatcher($container);
+		parent::__construct(self::APP_ID);
 	}
 
-	private function connectWatcher(IAppContainer $container) {
-		/** @var IRootFolder $root */
-		$root = $container->query(IRootFolder::class);
-		$root->listen('\OC\Files', 'postWrite', function (Node $node) use ($container) {
-			/** @var Watcher $watcher */
-			$watcher = $container->query(Watcher::class);
-			$watcher->postWrite($node);
-		});
+	public function register(IRegistrationContext $context): void {
+		$context->registerEventListener(NodeWrittenEvent::class, PostWriteListener::class);
+	}
+
+	public function boot(IBootContext $context): void {
 	}
 }