+++ lib/SizeHelper.php
@@ -1,11 +1,14 @@
 <?php
+
 declare(strict_types=1);
+
 /**
  * @copyright Copyright (c) 2017, Roeland Jago Douma <roeland@famdouma.nl>
  *
  * @author Roeland Jago Douma <roeland@famdouma.nl>
+ * @author Richard Steinmetz <richard@steinmetz.cloud>
  *
- * @license GNU AGPL version 3 or any later version
+ * @license AGPL-3.0-or-later
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Affero General Public License as
@@ -27,7 +30,10 @@ namespace OCA\PreviewGenerator;
 use OCP\IConfig;
 
 class SizeHelper {
-
+	/**
+	 * @param IConfig $config
+	 * @return int[][]
+	 */
 	public static function calculateSizes(IConfig $config): array {
 		/*
 		 * First calculate the systems max sizes
@@ -42,31 +48,30 @@ class SizeHelper {
 		$maxW = (int)$config->getSystemValue('preview_max_x', 4096);
 		$maxH = (int)$config->getSystemValue('preview_max_y', 4096);
 
-		$s = 32;
-		while($s <= $maxW || $s <= $maxH) {
+		$s = 64;
+		while ($s <= $maxW || $s <= $maxH) {
 			$sizes['square'][] = $s;
-			$s *= 2;
+			$s *= 4;
 		}
 
-		$w = 32;
-		while($w <= $maxW) {
+		$w = 64;
+		while ($w <= $maxW) {
 			$sizes['width'][] = $w;
-			$w *= 2;
+			$w *= 4;
 		}
 
-		$h = 32;
-		while($h <= $maxH) {
+		$h = 64;
+		while ($h <= $maxH) {
 			$sizes['height'][] = $h;
-			$h *= 2;
+			$h *= 4;
 		}
-
 
 		/*
 		 * Now calculate the user provided max sizes
-		 * Note that only powers of 2 matter but if users supply different
+		 * Note that only powers of 4 matter but if users supply different
 		 * stuff it is their own fault and we just ignore it
 		 */
-		$getCustomSizes = function(IConfig $config, $key) {
+		$getCustomSizes = function (IConfig $config, $key) {
 			$TXT = $config->getAppValue('previewgenerator', $key, '');
 			$values = [];
 			if ($TXT !== '') {