+++ CHANGELOG.md
@@ -1,3 +1,113 @@
+# Changelog
+
+## 5.2.4 - 2023-04-20
+### Fixed
+- Harden error handling of deleted files check
+
+## 5.2.3 - 2023-04-19
+### Changed
+- Skip deleted files
+
+## 5.2.2 - 2023-04-04
+### Changed
+- Fix PHP 8.2 deprecation warnings (thanks to @dotlambda)
+- Adjust requested preview sizes
+
+## 5.2.1 - 2023-02-24
+### Added
+- Support for PHP 7.4
+
+## 5.2.0 - 2023-02-10
+### Added
+- Support for Nextcloud 26
+- Support for PHP 8.2
+
+### Removed
+- Support for Nextcloud 23 (EOL)
+- Support for PHP 7.4 (EOL)
+
+### Changed
+- Modernize code
+
+## 5.1.1 - 2022-11-08
+### Fixed
+- Generate preview specifications once
+
+## 5.1.0 - 2022-09-23
+### Added
+- Nextcloud 25 support
+
+### Removed
+- Nextcloud 22 support
+
+## 5.0.0 - 2022-05-17
+### Added
+- Improve documentation around memory_limit (thanks to @proletarius101)
+- Improve documentation of preview:generate-all command
+
+### Changed
+- Updated maintainers
+
+### Removed
+- PHP 7.3 support
+- Nextcloud 21 support
+- preview:delete_old command
+
+## 4.0.0 - 2022-02-05
+### Added
+- PHP 8.1 support
+- Nextcloud 24 support
+
+### Removed
+- PHP 7.2 support
+- Nextcloud 20 support
+
+## 3.4.1 - 2021-12-15
+### Fixed
+- Dependency error on Nextcloud 20
+
+## 3.4.0 - 2021-12-06
+### Added
+- Skip external mounts with disabled previews (thanks to @hashworks)
+- Allow multiple paths and user ids (thanks to @hammer065)
+
+## 3.3.0 - 2021-11-30
+### Added
+- Nextcloud 23 support
+
+## 3.2.0 - 2021-11-09
+### Added
+- Nextcloud 22 support
+
+## 3.1.1 - 2021-01-27
+### Changed
+- Use new batch preview method
+
+### Fixed
+- Init filesystem for the users
+- Handle unavailable storages more graceful
+
+## 3.1.0 -2020-12-21
+### Added
+- Nextcloud 21 support
+
+## 3.0.0 - 2020-08-31
+### Added
+- Nextcloud 20 support
+- IBootstap support
+- Listen to new events hence only initialize the FS if needed
+
+### Removed
+- Nextcloud <20 support
+
+## 2.3.0 - 2020-04-03
+### Added
+- Nextcloud 19 support
+
+## 2.2.0 - 2019-12-05
+### Added
+- Nextcloud 18 support
+
 ## 2.1.0 - 2019-03-05
 ### Added
 - Nextcloud 16 support